// src/types/payload.ts
export interface ServiceProvider {
// ...
  caseStudies?:
    | {
        clientName?: string | null;
        title?: string | null;
        budget?: ('less_than_1k' | '1k_to_5k' | '5k_to_10k' | '10k_to_50k' | '50k_plus') | null;
        timeline?: ('less_than_1_month' | '1_to_3_months' | '3_to_6_months' | '6_to_12_months' | '1_year_plus') | null;
        /**
         * This field is automatically generated from the project name. eg. My Project -> my-project
         */
        caseStudySlug?: string | null;
        description?: string | null;
        projectUrl?: string | null;
        files?:
          | (
              | {
                  relationTo: 'file-uploads';
                  value: number | FileUpload;
                }
              | {
                  relationTo: 'remote-file-uploads';
                  value: number | RemoteFileUpload;
                }
            )[]
          | null;
        images?:
          | (
              | {
                  relationTo: 'image-uploads';
                  value: number | ImageUpload;
                }
              | {
                  relationTo: 'remote-file-uploads';
                  value: number | RemoteFileUpload;
                }
            )[]
          | null;
        serviceCategories?: {
          /**
           * Select the service categories this case study belongs to
           */
          categories?: (number | Category)[] | null;
          /**
           * Select the sub-categories this case study belongs to
           */
          subCategories?: (number | SubCategory)[] | null;
        };
        achievedMetrics?:
          | {
              metric?: string | null;
              id?: string | null;
            }[]
          | null;
        id?: string | null;
      }[]
    | null;
// ...
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "image-uploads".
 */
export interface ImageUpload {
  id: number;
  /**
   * The service provider this image belongs to
   */
  serviceProvider?: (number | null) | ServiceProvider;
  /**
   * Optional description of the image
   */
  description?: string | null;
  updatedAt: string;
  createdAt: string;
  url?: string | null;
  thumbnailURL?: string | null;
  filename?: string | null;
  mimeType?: string | null;
  filesize?: number | null;
  width?: number | null;
  height?: number | null;
  focalX?: number | null;
  focalY?: number | null;
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "remote-file-uploads".
 */
export interface RemoteFileUpload {
  id: number;
  /**
   * Type of the remote file
   */
  type: 'image' | 'document';
  /**
   * The service provider this file is associated with
   */
  serviceProvider?: (number | null) | ServiceProvider;
  /**
   * Remote file URL from our remote storage
   */
  remoteURL: string;
  /**
   * Optional description of the file
   */
  description?: string | null;
  updatedAt: string;
  createdAt: string;
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "file-uploads".
 */
export interface FileUpload {
  id: number;
  /**
   * The service provider this file belongs to. Used during registration process for service provider.
   */
  serviceProvider?: (number | null) | ServiceProvider;
  /**
   * Optional description of the file
   */
  description?: string | null;
  updatedAt: string;
  createdAt: string;
  url?: string | null;
  thumbnailURL?: string | null;
  filename?: string | null;
  mimeType?: string | null;
  filesize?: number | null;
  width?: number | null;
  height?: number | null;
  focalX?: number | null;
  focalY?: number | null;
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "categories".
 */
export interface Category {
  id: number;
  name: string;
  /**
   * URL friendly version of a name used to identify a category
   */
  slug?: string | null;
  /**
   * When a category is displayed, this icon will be used
   */
  icon: string;
  subCategories?: (number | SubCategory)[] | null;
  /**
   * A short description of the category
   */
  description?: {
    root: {
      type: string;
      children: {
        type: string;
        version: number;
        [k: string]: unknown;
      }[];
      direction: ('ltr' | 'rtl') | null;
      format: 'left' | 'start' | 'center' | 'right' | 'end' | 'justify' | '';
      indent: number;
      version: number;
    };
    [k: string]: unknown;
  } | null;
  updatedAt: string;
  createdAt: string;
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "sub-categories".
 */
export interface SubCategory {
  id: number;
  name: string;
  /**
   * URL friendly version of a name used to identify a category
   */
  slug?: string | null;
  parentCategory: number | Category;
  updatedAt: string;
  createdAt: string;
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "team-members".
 */
export interface TeamMember {
  id: number;
  name: string;
  position: string;
  image?: (number | null) | ImageUpload;
  updatedAt: string;
  createdAt: string;
}